Added some op code functionality
Also messed with the memory accessor such that the same function can be used to return the value of varying amount of hex strings. This is helpful so the same function could be used to read a constant, or a memory index. 

I've also started telling my OS how to write an OS so it can do the project for me. Unfortunately i also gave it free will and it doesn't agree with me not doing my own work. But if I write the OS that writes the OS, is it actually my own work? Too philosophical for me to think about. And I'm staring to give someone too many ideas for a nasty iProject 5.